pyproject.toml
[tool.poetry]
name = "forum"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.9"
django = "^5.1.1"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
djoser = "^2.2.3"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.4.0"
gunicorn = "^23.0.0"
celery = "^5.4.0"
redis = "^5.1.1"
flower = "^2.0.1"
django-filter = "^24.3"
django-simple-history = "^3.7.0"
channels = "^4.1.0"
daphne = "^4.1.2"
pymongo = "^4.10.1"
punq = "^0.7.0"
bleach = "^6.1.0"
django-cryptography = "^1.1"
django-ratelimit = "^4.1.0"
drf-yasg = "^1.21.8"
django-redis = "^5.4.0"

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.1"
pylint-django = "^2.5.5"
ipython = "^8.28.0"
bandit = "^1.7.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

